Implement ADS1115 analog to digital converter driver
Use single shot conversions to preserve power.

+++ power.hxx
@@ -14,8 +14,8 @@ public:
     Power();
     ~Power();
 
-    unsigned int voltage(void) const;
-    unsigned int current(void) const;
+    float voltage(void) const;
+    float current(void) const;
 
 };
